auth gate

+++ app/Http/Controllers/QuestionsController.php
@@ -71,6 +71,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+        }
+
         return view("questions.edit", compact('question'));
     }
 
@@ -83,6 +87,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+        }
+
         $question->update($request->only('title','body'));
 
         return redirect('/questions')->with('success','Questions Updated successfuly');
@@ -96,6 +105,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access Denied");
+        }
+
         $question->delete();
 
         return redirect('/questions')->with('success','Question deleted');
